+++ src/main/java/com/hoa/shopbanhang/application/services/IProductService.java
@@ -20,7 +20,7 @@ public interface IProductService {
 
   Product updateProduct(UpdateProductInput updateProductInput);
 
-  void updateStockProduct(Long productId, Long amount, Boolean isBuy);
+  void updateStockProduct(Long productId, Integer amount, Boolean isBuy);
 
   RequestResponse deleteById(Long id);
 }
